renames the TextareaInput files according to coding conventions
